Add option to only include a feed item if its page has a date
Add option --require-dates to only include a feed item if the source
page has one of the parseable date formats (from meta tags). This is a
nice way to remove non-relevant results from blog posts or news sites,
but also requires the script to load every link (--follow), which is
slow and generates excessive traffic to the source site.