Allow using XPath expressions to select publishing dates
Add support for XPath expressions, so they can be used to select text
elements containing the data a page was last updated, and then parsing
it. The process uses three options, --date-from-xpath, --xpath-regex,
and --xpath-date-fmt, working as a pipeline. The first gets the text,
the second optionally selects a substring from it using a regular
expression, and the third defines the date/time format to parse it.

This adds a dependency to library lxml and a lot of extra complexity,
but sometimes there is no easier way to grab this date.