Allow use of CSS Selectors when picking the document body
There are a lot of cases where the XPath expressions required to select
the document body from some sources are too complex or just confusing.
This commit adds CSS Selectors as an alternative way to get the same
result, while keeping the support for XPath.

Next step is adding support for CSSS in other places too (dates,
element removal, ...)