Allow use of CSS Selectors for picking dates
Add new options --date-from-csss, --csss-date-regex, and --csss-date-fmt
to select the elements with publishing dates from the document using CSS
Selectors. This works as the equivalent XPath options, but with a
simpler and cleaner syntax and it is restricted to the text content of
the elements (i.e. there is no way to pick dates from attributes).