Refactor item date finding code
Move the code that finds the item publishing date to a separate function
and reuse it between the two calls. This also centralizes all document
tree parsing in a single call, at the cost of removing a small
optimization that can avoid the need of parsing at all (but only in very
limited situations).