Change priority of publish date finding options
Change the priorities of the options that find the publishing dates
of an item. Idea is to try first the ones that were explicitly listed
in command line and than the implicit sources, with options that can
return better date information having priority.

There is a bit of experimentation going on so new changes may follow.