Allow requesting documents in a particular language
Read preferred language(s) from command line options of LANG environment
variable and set the Accept-Language HTTP header accordingly. This
should work for web sites that change date formats, etc. according to
the languages sent in the browser options.

Using the environment language is a two-edged sword: it can help users as
usually they debug in the same system where newslinkrss runs and the
typical case is to have the same language settings for the browser used
for debugging and the environment) but can also introduce a new dependency
on external configuration that is not obvious at first.