Add option to set user-defined cookies for requests
Add a new "--cookie" command line option to allow the user to define
cookies that will be sent to the server in every request. This helps in
some specific situations like choosing the correct geographical variant
of sites that customize news based on it. It also plays well with option
"--no-cookies", allowing creation of "read-only" cookies.