Add options to capture item authors from DOM
Add new options --author-from-xpath and --author-from-csss and auxiliary
options --xpath-author-regex and --csss-author-regex to extract author
information from the page elements. This may be useful as it allows
filtering authors (in the news reader) from sites that do not include
author in metadata. newslinkrss stills try to get author information
automatically if this options are not given or if they do not return any
useful information.